+++ src/GlobalExceptionHandler/WebApi/ExceptionHandlerConfiguration.cs
@@ -54,28 +54,4 @@ namespace GlobalExceptionHandler.WebApi
 			};
 		}
 	}
-
-	class ExceptionTypePolymorphicComparer : IComparer<Type>
-	{
-		public int Compare(Type x, Type y)
-		{
-			var depthOfX = 0;
-			var currentType = x;
-			while (currentType != typeof(object))
-			{
-				currentType = currentType.BaseType;
-				depthOfX++;
-			}
-
-			var depthOfY = 0;
-			currentType = y;
-			while (currentType != typeof(object))
-			{
-				currentType = currentType.BaseType;
-				depthOfY++;
-			}
-
-			return depthOfY.CompareTo(depthOfX);
-		}
-	}
 }
